Remove shotover specific logic from DockerCompose
