+++ shotover-proxy/examples/cassandra_bench.rs
@@ -1,5 +1,5 @@
 use clap::Parser;
-use test_helpers::docker_compose::DockerCompose;
+use test_helpers::docker_compose::docker_compose;
 use test_helpers::latte::Latte;
 use test_helpers::shotover_process::ShotoverProcessBuilder;
 
@@ -25,7 +25,7 @@ async fn main() {
     let latte = Latte::new(args.rate, 1);
     let bench = "read";
     {
-        let _compose = DockerCompose::new(&format!("{}/docker-compose.yaml", args.config_dir));
+        let _compose = docker_compose(&format!("{}/docker-compose.yaml", args.config_dir));
 
         let shotover = ShotoverProcessBuilder::new_with_topology(&format!(
             "{}/topology.yaml",
