+++ shotover-proxy/examples/cassandra_cluster_bench.rs
@@ -1,4 +1,4 @@
-use test_helpers::docker_compose::DockerCompose;
+use test_helpers::docker_compose::docker_compose;
 use test_helpers::latte::Latte;
 use test_helpers::shotover_process::ShotoverProcessBuilder;
 
@@ -10,7 +10,7 @@ async fn main() {
     let config_dir = "example-configs/cassandra-cluster-v4";
     let bench = "read";
     {
-        let _compose = DockerCompose::new(&format!("{}/docker-compose.yaml", config_dir));
+        let _compose = docker_compose(&format!("{}/docker-compose.yaml", config_dir));
         let shotover =
             ShotoverProcessBuilder::new_with_topology(&format!("{}/topology.yaml", config_dir))
                 .start()
