+++ shotover-proxy/tests/kafka_int_tests/mod.rs
@@ -1,5 +1,6 @@
 use serial_test::serial;
-use test_helpers::{docker_compose::DockerCompose, shotover_process::ShotoverProcessBuilder};
+use test_helpers::docker_compose::docker_compose;
+use test_helpers::shotover_process::ShotoverProcessBuilder;
 
 mod test_cases;
 
@@ -7,7 +8,7 @@ mod test_cases;
 #[serial]
 async fn passthrough_standard() {
     let _docker_compose =
-        DockerCompose::new("tests/test-configs/kafka/passthrough/docker-compose.yaml");
+        docker_compose("tests/test-configs/kafka/passthrough/docker-compose.yaml");
     let shotover = ShotoverProcessBuilder::new_with_topology(
         "tests/test-configs/kafka/passthrough/topology.yaml",
     )
@@ -23,7 +24,7 @@ async fn passthrough_standard() {
 #[serial]
 async fn passthrough_encode() {
     let _docker_compose =
-        DockerCompose::new("tests/test-configs/kafka/passthrough/docker-compose.yaml");
+        docker_compose("tests/test-configs/kafka/passthrough/docker-compose.yaml");
     let shotover = ShotoverProcessBuilder::new_with_topology(
         "tests/test-configs/kafka/passthrough/topology-encode.yaml",
     )
